refactor appendUserInfo to a utility function
